New interactor: UpdateLibraryShows
This is a merge of UpdateFollowedShows and
UpdateWatchedShows. There's no need to keep them
separate now, as they are interlinked.
