--- domain/src/main/java/app/tivi/domain/interactors/UpdateWatchedShows.kt
@@ -1,58 +1,0 @@
-/*
- * Copyright 2019 Google LLC
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package app.tivi.domain.interactors
-
-import app.tivi.data.showimages.ShowImagesStore
-import app.tivi.data.shows.ShowStore
-import app.tivi.data.util.fetch
-import app.tivi.data.watchedshows.WatchedShowsStore
-import app.tivi.domain.Interactor
-import app.tivi.util.AppCoroutineDispatchers
-import app.tivi.util.Logger
-import app.tivi.util.parallelForEach
-import kotlinx.coroutines.ensureActive
-import kotlinx.coroutines.withContext
-import me.tatarka.inject.annotations.Inject
-
-@Inject
-class UpdateWatchedShows(
-    private val watchedShowsStore: WatchedShowsStore,
-    private val showsStore: ShowStore,
-    private val showImagesStore: ShowImagesStore,
-    private val dispatchers: AppCoroutineDispatchers,
-    private val logger: Logger,
-) : Interactor<UpdateWatchedShows.Params>() {
-    override suspend fun doWork(params: Params) {
-        withContext(dispatchers.io) {
-            val shows = watchedShowsStore.fetch(Unit, params.forceRefresh)
-
-            shows.parallelForEach { show ->
-                ensureActive()
-                showsStore.fetch(show.showId)
-
-                ensureActive()
-                try {
-                    showImagesStore.fetch(show.showId)
-                } catch (t: Throwable) {
-                    logger.e(t, "Error while fetching images for show: ${show.showId}")
-                }
-            }
-        }
-    }
-
-    data class Params(val forceRefresh: Boolean)
-}
